feat 로그인필터 구현 및 전반적인 코드 개선 및 주석 추가 및 사용자 환전 전체 조회 기능 개선
필요하지 않은 import문 제거, 중복이라고 생각하는 코드 메소드로 추출, 요청이 취소되면 사용자가 요청한 횟수가 감소되지 않은 문제 해결

+++ src/main/java/com/sparta/currency_user/controller/UserController.java
@@ -12,6 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import javax.naming.AuthenticationException;
 import java.util.List;
 
 @RestController
@@ -21,41 +22,45 @@ public class UserController {
 
     private final UserService userService;
 
-    @GetMapping
+    @GetMapping // 전체 사용자 조회
     public ResponseEntity<List<UserResponseDto>> findUsers() {
         return ResponseEntity.ok().body(userService.findAll());
     }
 
-    @GetMapping("/{id}")
+    @GetMapping("/{id}") // 단건 사용자 조회
     public ResponseEntity<UserResponseDto> findUser(@PathVariable Long id) {
         return ResponseEntity.ok().body(userService.findById(id));
     }
 
-    @PostMapping
+    @PostMapping("signup") // 회원 가입
     public ResponseEntity<UserResponseDto> createUser(@Validated @RequestBody UserRequestDto userRequestDto) {
         return ResponseEntity.ok().body(userService.save(userRequestDto));
     }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable Long id,HttpServletRequest request) {
+    @DeleteMapping("/{id}") // 로그인한 사용자 id와 요청한 id가 일치할 때 삭제
+    public ResponseEntity<String> deleteUser(@PathVariable Long id, HttpServletRequest request) {
         HttpSession session = request.getSession(false);
-        if(session!=null){
-            session.invalidate();
+        Long sessionKey = (Long) session.getAttribute("SESSION_KEY");
+
+        if (!sessionKey.equals(id)) { // 일치 하지않으면
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("권한이 없습니다.");
         }
-        userService.deleteUserById(id);
+        userService.deleteUserById(id); // 일치하면 삭제
         return ResponseEntity.ok().body("정상적으로 삭제되었습니다.");
     }
-    @PostMapping("/login")
+
+    @PostMapping("/login") // 로그인
     public ResponseEntity<Void> login(@Validated @RequestBody UserRequestDto requestDto, HttpServletRequest request) {
         User user = userService.login(requestDto);
         HttpSession session = request.getSession();
-        session.setAttribute("SESSION_KEY",user.getId());
+        session.setAttribute("SESSION_KEY", user.getId());
         return new ResponseEntity<>(HttpStatus.OK);
     }
-    @DeleteMapping("/logout")
+
+    @DeleteMapping("/logout") //로그아웃
     public ResponseEntity<Void> logout(HttpServletRequest request) {
         HttpSession session = request.getSession(false);
-        if(session!=null){
+        if (session != null) {
             session.invalidate();
         }
         return new ResponseEntity<>(HttpStatus.OK);
